requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.36
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.127.0
apscheduler>=3.10.4
python-dotenv>=1.0.1
dateparser>=1.2.0
python-dateutil>=2.9.0
beautifulsoup4>=4.12.3
lxml>=5.2.1
httpx>=0.27.0
